Refactor trivial expression substitution into separate file
